Upstream version 2.2.0~pre1

Conflicts:
	Makefile.in
	bin/Makefile.in
	compat/Makefile.in
	config.h.in
	configure
	doc/Makefile.in
	lib/Makefile.in
	m4/Makefile.in
	modules/Makefile.in
	modules/access_output/Makefile.in
	modules/audio_filter/Makefile.in
	modules/control/Makefile.in
	modules/gui/Makefile.in
	modules/gui/ios_dialog_provider/Makefile.in
	modules/gui/macosx/Makefile.in
	modules/gui/macosx_dialog_provider/Makefile.in
	modules/gui/minimal_macosx/Makefile.in
	modules/gui/qt4/Makefile.in
	modules/gui/skins2/Makefile.in
	modules/hw/vdpau/Makefile.in
	modules/misc/Makefile.in
	modules/mux/Makefile.in
	modules/packetizer/Makefile.in
	modules/stream_out/Makefile.in
	modules/text_renderer/Makefile.in
	modules/video_filter/Makefile.in
	modules/video_output/Makefile.in
	modules/visualization/Makefile.in
	share/Makefile.in
	src/Makefile.in
	test/Makefile.in

+++ modules/gui/macosx/about.h
@@ -2,7 +2,7 @@
  * about.h: MacOS X About Panel
  *****************************************************************************
  * Copyright (C) 2001-2013 VLC authors and VideoLAN
- * $Id: 34f1c6aa67fc9863eb6bac0a2ad944c38e6c0f1e $
+ * $Id: 8bab02e484f7e96bc733ce5cae2e7c8d8a2271c2 $
  *
  * Authors: Derk-Jan Hartman <thedj@users.sourceforge.net>
  *          Felix Paul Kühne <fkuehne -at- videolan.org>
@@ -46,7 +46,7 @@
 
     NSTimer *o_scroll_timer;
     float f_current;
-    float f_end;
+    CGFloat f_end;
     NSTimeInterval i_start;
     BOOL b_restart;
     BOOL b_isSetUp;
